feat(analytics): streaming Suspense + layout-matched skeleton + error boundary
Make the analytics page feel "Amazon-smooth" by streaming the heavy
SLA-metrics computation behind Suspense instead of blocking the whole
route. Three layered improvements:

1. AnalyticsSkeleton — a structurally-matched skeleton (KPI grid,
   gauges, charts, distribution rows) replaces the generic dashboard
   skeleton, so the placeholder you see during filter changes is
   visually consistent with the real content.

2. page.tsx is now a thin shell that renders the gradient header +
   filter bar instantly (only fast Team/Service/User lookups), then
   defers calculateSLAMetrics into <AnalyticsContent> wrapped in a
   keyed <Suspense>. The key is a hash of all filters, so each
   filter change shows the skeleton again while the new data loads
   rather than freezing on stale numbers.

3. analytics/error.tsx — route-level error boundary so a single
   failing query degrades to a retry card instead of blowing up the
   whole app and showing "Application Error / Something went wrong".

Combined effect: filter clicks now feel instant (shell stays put,
skeleton swaps in), and a transient DB hiccup on one query degrades
gracefully instead of unmounting the entire page.
